Themed code-of-conduct and collaborate-with-us. Transferred content of coc to code-of-conduct and deleted coc to replicate paths from v1
